+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added experimental query parameters builder `ydb.ParamsBuilder()`
+
 ## v3.56.2
 * Fixed return private error for commit to stopped partition in topic reader.
 * Stopped wrapping err error as transport error at topic streams (internals)